[ISSUE-17] - remove module onboarding and move splash for module app
