+++ app/src/main/AndroidManifest.xml
@@ -9,13 +9,12 @@
         android:roundIcon="@mipmap/ic_netflix_round"
         android:label="@string/app_name"
         android:supportsRtl="true"
-        android:theme="@style/Theme.StreamPlayerApp"
-        android:name=".CustomApplication">
+        android:name="com.codandotv.streamplayerapp.CustomApplication">
 
         <activity
-            android:name=".MainActivity"
+            android:name="com.codandotv.streamplayerapp.MainActivity"
             android:exported="true"
-            android:theme="@android:style/Theme.Translucent.NoTitleBar">
+            android:theme="@style/SplashScreenTheme">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
