+++ app/src/main/java/com/codandotv/streamplayerapp/MainActivity.kt
@@ -1,13 +1,16 @@
 package com.codandotv.streamplayerapp
 
-import android.content.Intent
 import android.os.Bundle
 import androidx.activity.ComponentActivity
-import com.codandotv.streamplayerapp.feature_onboarding.OnboardingActivity
+import androidx.activity.compose.setContent
+import com.codandotv.streamplayerapp.screens.SplashScreen
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        startActivity(Intent(this@MainActivity, OnboardingActivity::class.java))
+
+        setContent{
+            SplashScreen()
+        }
     }
 }
