+++ app/src/main/java/com/codandotv/streamplayerapp/screens/SplashScreen.kt
@@ -1,11 +1,9 @@
-package com.codandotv.streamplayerapp.feature_onboarding.screens
+package com.codandotv.streamplayerapp.screens
 
 import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material3.MaterialTheme
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.Alignment
@@ -16,7 +14,7 @@ import com.airbnb.lottie.compose.LottieAnimation
 import com.airbnb.lottie.compose.LottieCompositionSpec
 import com.airbnb.lottie.compose.animateLottieCompositionAsState
 import com.airbnb.lottie.compose.rememberLottieComposition
-import com.codandotv.streamplayerapp.feature_onboarding.R
+import com.codandotv.streamplayerapp.core_shared_ui.R
 
 @Composable
 fun SplashScreen() {
