+++ build.gradle.kts
@@ -18,6 +18,7 @@ allprojects {
         google()
         mavenCentral()
         maven(url = uri("https://oss.sonatype.org/content/repositories/snapshots/"))
+        
     }
     configureAndroid()
 }
