+++ core-shared-ui/src/main/res/raw/logo.json
