--- feature-onboarding/src/main/AndroidManifest.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<manifest xmlns:android="http://schemas.android.com/apk/res/android">
-
-    <application>
-        <activity
-            android:name=".OnboardingActivity"
-            android:exported="false"
-            android:theme="@style/SplashScreenTheme"/>
-    </application>
-
-</manifest>
