--- feature-onboarding/src/main/res/values/strings.xml
@@ -1,1 +1,0 @@
-<resources></resources>
