+++ settings.gradle.kts
@@ -1,5 +1,4 @@
 include(":app")
 include(":feature-list-streams")
 include(":core-networking")
-include(":feature-onboarding")
 include(":core-shared-ui")
